Comic2requirements.pip.txt
Pillow==2.4.0
argparse==1.2.1
wsgiref==0.1.2
